fix: use hooks
