+++ packages/web/src/components/react/hooks/useDemoMode.tsx
@@ -1,0 +1,29 @@
+import { useEffect, useRef } from "react";
+import { useSearchParams } from "react-router-dom";
+import { useContentTypeSeeder } from "./useContentTypeSeeder";
+import { useSpaceSeeder } from "./useSpaceSeeder";
+
+// Hook to handle demo mode logic
+export function useDemoMode() {
+	const [searchParams] = useSearchParams();
+	const isDemoMode = searchParams.get("demo") === "1";
+	const isSeededRef = useRef(false);
+
+	const { seedSpaces } = useSpaceSeeder();
+	const { seedContentTypes } = useContentTypeSeeder();
+
+	useEffect(() => {
+		if (isDemoMode && !isSeededRef.current) {
+			console.log("Demo mode enabled - seeding demo data");
+			isSeededRef.current = true;
+
+			// Seed spaces first, then content types
+			seedSpaces();
+			seedContentTypes();
+		}
+	}, [isDemoMode, seedSpaces, seedContentTypes]);
+
+	return {
+		isDemoMode,
+	};
+}
